add custom filters for date and booleans
